# Improve the perfomance of update products search vector task