I want to merge this change because it should cover: https://github.com/saleor/saleor/issues/15180

Internal link: https://linear.app/saleor/issue/MERX-67/optimize-update-products-search-vector-task

The results:
Tested on main, and on this branch

|   | Main |  This branch | 
| ------------- | ------------- | -- |
| Exectuion time  | 206.42s  | 27.04s | 
| DB queries called  | 9000+  |  43 |
| *Maximum memory usage: |  1658.36MB |  766.98MB | 

`*` memory usage tested by `memory_profiler` connected to celery worker 👇 
```python
from memory_profiler import memory_usage
mem_usage = memory_usage(7989, interval=.2, timeout=220)
max(mem_usage)
1658.359375
```

> [!NOTE]  
> 27 sec is still a big amount, but I had a really big dataset of attributes with rich_test and plain_text. The bulk_update took more than 20second on my setup.

<!-- Please mention all relevant issue numbers. -->

# Impact

- [ ] New migrations
- [ ] New/Updated API fields or mutations
- [ ] Deprecated API fields or mutations
- [ ] Removed API types, fields, or mutations

# Docs

<!-- Docs are stored in a separate repository: https://github.com/saleor/saleor-docs/. -->
<!-- Please provide a link to the PR that updates documentation for your changes. -->
<!-- If changes in docs are not required, please mention that in the description. -->

- [ ] Link to documentation:

# Pull Request Checklist

<!-- Please keep this section. It will make the maintainer's life easier. -->

- [ ] Privileged queries and mutations are either absent or guarded by proper permission checks
- [ ] Database queries are optimized and the number of queries is constant
- [ ] Database migrations are either absent or optimized for zero downtime
- [ ] The changes are covered by test cases
- [ ] All new fields/inputs/mutations have proper labels added (`ADDED_IN_X`, `PREVIEW_FEATURE`, etc.)
- [ ] All migrations have proper dependencies
- [ ] All indexes are added concurrently in migrations
- [ ] All RunSql and RunPython migrations have revert option defined