@aniav 2024-04-05
@saleor/merchant-experience FYI the mentioned @sentry-io issues are indeed happening ☝️